+++ src/LCT.APICommunications/RetryHttpClientHandler.cs
@@ -43,5 +43,24 @@ namespace LCT.APICommunications
                 return await base.SendAsync(request, cancellationToken); // Pass the request to the next handler (HttpClient)
             });
         }
+        public static async Task ExecuteWithRetryAsync(Func<Task> action)
+        {
+            var retryPolicy = Policy
+                .Handle<WebException>()
+                .WaitAndRetryAsync(ApiConstant.APIRetryCount, // Retry up to 3 times
+                    attempt => attempt switch
+                    {
+                        1 => TimeSpan.FromSeconds(ApiConstant.APIRetryIntervalFirst),  // 1st retry after 5 seconds
+                        2 => TimeSpan.FromSeconds(ApiConstant.APIRetryIntervalSecond), // 2nd retry after 10 seconds
+                        3 => TimeSpan.FromSeconds(ApiConstant.APIRetryIntervalThird), // 3rd retry after 30 seconds
+                        _ => TimeSpan.Zero // Default (not used)
+                    },
+                    onRetry: (exception, timespan, attempt, context) =>
+                    {
+                        Logger.Debug($"Retry attempt {attempt} due to: {exception?.Message ?? "No exception"}");
+                    });
+
+            await retryPolicy.ExecuteAsync(action);
+        }
     }
 }
